모든 ID 참조 PlayerIdManager로 변경

+++ Assets/0_ColorRandomDefance/1_Script/1_Unit/Multi_TeamSoldier.cs
@@ -51,7 +51,7 @@ public class Multi_TeamSoldier : MonoBehaviourPun
     private ObstacleAvoidanceType originObstacleType;
     protected Animator animator;
     protected RPCable rpcable;
-    public int UsingID => rpcable.UsingId;
+    public byte UsingID => (byte)rpcable.UsingId;
     [SerializeField] protected EffectSoundType normalAttackSound;
     public float normalAttakc_AudioDelay;
